Updated map topology management
