+++ py-modules/map-topology/macrostrat/map_topology/tests/test_map_topology.py
@@ -5,7 +5,8 @@ from shapely.geometry import Point
 
 from macrostrat.database import Database
 from macrostrat.database.utils import template_database
-from macrostrat.map_topology import create_topo_context, update_maps
+from macrostrat.map_topology import create_topo_context
+from macrostrat.map_topology.update import update_maps
 
 
 def geom(_shape, srid=4326):
